Graphics.Plotly

Description

Plotly Plots in Haskell

Documentation

plot

Arguments

:: FilePath
Target path to .html file

-> Script
Plotly Script

-> IO ()

Save Plot to HTML File

scatter'

Arguments

:: [String]
Trace Names

-> [[Double]]
xs

-> [[Double]]
ys

-> PlotConfig
Plot Config

-> Script
Plotly Script

Scatter Plot with individual x values per trace

scatter

Arguments

:: [String]
Trace Names

-> [Double]
xs

-> [[Double]]
ys

-> PlotConfig
Plot Config

-> Script
Plotly Script

Scatter Plot with same x axis for all ys

histogram

Arguments

:: [String]
Trace Names

-> [[Double]]
Traces

-> PlotConfig
Plot Config

-> Script
Plotly Script

Histogram Plot

heatmap

Arguments

:: [String]
X Categories

-> [String]
Y Categories

-> [[Double]]
Data Matrix

-> PlotConfig
Plot Config

-> Script
Plotly Script

Heatmap Plot

scatter3d

Arguments

:: [String]
Trace Names

-> [[Double]]
xs

-> [[Double]]
ys

-> [[Double]]
zs

-> PlotConfig
Plot Config

-> Script
Plotly Script

3D Scatter Plot

surface''

Arguments

:: [String]
Trace Names

-> [[[Double]]]
zs

-> PlotConfig
Plot Config

-> Script
Ploty Script

3D Surface Plot where X and Y are indices

surface'

Arguments

:: [String]
Trace Names

-> [Double]
xs

-> [Double]

-> [[[Double]]]
zs

-> PlotConfig
Plot Config

-> Script
Plotly Script

3D Surface Plot with same xs and ys for all zs

surface

Arguments

:: [String]
Trace Names

-> [[Double]]
xs

-> [[Double]]
ys

-> [[[Double]]]
zs

-> PlotConfig
P

-> Script
Trace

3D Surface Plot

parcoord'

Arguments

:: [String]
Axis Names

-> [[[Double]]]
traces ( rows ( values ) )

-> PlotConfig
Plot Config

-> Script
Plot Script

Parallel Coordinate Plot without line gradient

parcoord

Arguments

:: [Double]
Color column

-> [String]
Axis Labels

-> [[[Double]]]
traces ( rows ( values ) )

-> PlotConfig
Plot Config

-> Script
Plotly Script

Parallel Coordinate Plot with line gradient

type Script = ByteString

Internal Type Alias

data Mode

Scatter Mode

Constructors

Lines
Draw Lines

Markers
Draw Markers only

LinesMarkers
Draw Both

data Type

Plot Type

Constructors

Scatter
Scatter Plot

Histogram
Histogram

Heatmap
Heatmap

Scatter3D
3D Scatter

Surface
3D Surface

ParCoords
Parallel Coordinate Plot

data Layout

Layout for Plots

Constructors

Layout

Fields

title :: !String

Plot Title

xaxis :: !(Maybe Axis)

X Axis Layout

yaxis :: !(Maybe Axis)

Y Axis Layout

barmode :: !(Maybe BarMode)

Bar Chart Layout

height :: !Int

Plot Height in Pixels

width :: !Int

Plot Width in Pixels

showlegend :: !(Maybe Bool)

Show a Legend

margin :: !(Maybe Margin)

Margins

data BarMode

Bar Layout

Constructors

Stack
Stacked

Group
Grouped

Overlay
Overlayed

Relative
Relative

data AxisType

Axis Type

Constructors

Linear
Linear Axis (default)

Log
Logarithmic Axis

Date
Date Axis

Category
Categorical Axis

MultiCategory
Multi Category Axis (?)

data ColorScale

Color scales for Heamap / 3D Plots

Constructors

Blackbody

Bluered

Blues

Cividis

Earth

Electric

Greens

Greys

Hot

Jet

Picnic

Portland

Rainbow

RdBu

Reds

Viridis

YlGnBu

YlOrRd

data Symbol

Marker Symbols

Constructors

Circle

CircleOpen

Cross

Diamond

DiamondOpen

Square

SquareOpen

X

data Line

Parallel Coordinate Lines

Constructors

Line

Fields

showscale :: !(Maybe Bool)

Show Color Bar

reversescale :: !(Maybe Bool)

Reverse the Color Scale

colorscale :: !(Maybe ColorScale)

Color Scale

color :: ![Double]

Values represnting the color

data Margin

Layout Margins

Constructors

Margin

Fields

l :: !Int

Left Margin

r :: !Int

Right Margin

b :: !Int

Bottom Margin

t :: !Int

Top Margin

data PlotConfig

Trace Independent Plot Config

Constructors

PlotConfig

Fields

title' :: !String

Plot Title (default "")

xlabel :: !String

X Axis Label (default "")

ylabel :: !String

Y Axis Label (default "")

xmode :: !AxisType

X Axis Type (default Linear)

ymode :: !AxisType

Y Axis Type (default Linear)

barMode :: !BarMode

Bar Mode (default Stack)

lineMode :: !Mode

Line Mode (default Linear)

colorScale :: !ColorScale

Color Scale for 3D (default Viridis)

showScale :: !Bool

Show Color Bar (default True)

reverseScale :: !Bool

Reverse the color scale (default False)

legend :: !Bool

Show Legend (default True)

height :: !Int

Canvas Height in Pixels (default 800)

width :: !Int

Canvas Width in Pixels (default 800)

margin :: !Margin

Layout margin (default 66 66 66 66)

marker :: !Marker

Marker (default 0.5 Circle 1.0)

bins :: !XBins

Histogram bins (default 0.05 Nothing Nothing)

defaultConfig :: PlotConfig

Default PlotConfig