Graphics.Plotly.Internal

Description

Internal Functions and Types for Plotly Plots in Haskell

Documentation

fromConfig :: PlotConfig -> Layout

Config to Layout

type Script = ByteString

Internal Type Alias

showMaybe :: Show a => Maybe a -> String

Prints "null" for Nothing, works good with JSON

toHtml :: Script -> ByteString

Takes a Script and inserts it into the default HTML template

data Margin

Layout Margins

Constructors

Margin

Fields

l :: !Int

Left Margin

r :: !Int

Right Margin

b :: !Int

Bottom Margin

t :: !Int

Top Margin

data Layout

Layout for Plots

Constructors

Layout

Fields

title :: !String

Plot Title

xaxis :: !(Maybe Axis)

X Axis Layout

yaxis :: !(Maybe Axis)

Y Axis Layout

barmode :: !(Maybe BarMode)

Bar Chart Layout

height :: !Int

Plot Height in Pixels

width :: !Int

Plot Width in Pixels

showlegend :: !(Maybe Bool)

Show a Legend

margin :: !(Maybe Margin)

Margins

data AxisType

Axis Type

Constructors

Linear
Linear Axis (default)

Log
Logarithmic Axis

Date
Date Axis

Category
Categorical Axis

MultiCategory
Multi Category Axis (?)

data Axis

Axis Layout

Constructors

Axis

Fields

title :: !String

Axis Label

showGrid :: !Bool

Enable Grid

zeroLine :: !Bool

Enable Zero Line

type' :: !AxisType

Axis Type

data Marker

Marker Config

Constructors

Marker

Fields

size :: !Double

Marker Size

symbol :: !Symbol

Marker Symbol

opacity :: !Double

Marker Opacity

data Trace

Data Trace

Constructors

Trace

Fields

name :: !(Maybe String)

Trace Name

x :: ![Double]

X axis values

y :: !(Maybe [Double])

Y axis values

z :: !(Maybe [Double])

Z axis values

mode :: !(Maybe Mode)

Scatter Mode

type' :: !Type

Plot Type

marker :: !(Maybe Marker)

Marker Configuration

data TraceH

Heatmap Trace

Constructors

TraceH

Fields

z :: ![[Double]]

Data Matrix

x :: !(Maybe [String])

Y labels

y :: !(Maybe [String])

Y Labels

type' :: !Type

Heatmap

hoverOnGaps :: !(Maybe Bool)

Hover on Gaps

showScale :: !(Maybe Bool)

Show scale Bar

colorScale :: !(Maybe ColorScale)

Color Scale

data Type

Plot Type

Constructors

Scatter
Scatter Plot

Histogram
Histogram

Heatmap
Heatmap

Scatter3D
3D Scatter

Surface
3D Surface

data Mode

Scatter Mode

Constructors

Lines
Draw Lines

Markers
Draw Markers only

LinesMarkers
Draw Both

data Symbol

Marker Symbols

Constructors

Circle

CircleOpen

Cross

Diamond

DiamondOpen

Square

SquareOpen

X

data ColorScale

Color scales for Heamap / 3D Plots

Constructors

Blackbody

Bluered

Blues

Cividis

Earth

Electric

Greens

Greys

Hot

Jet

Picnic

Portland

Rainbow

RdBu

Reds

Viridis

YlGnBu

YlOrRd

data BarMode

Bar Layout

Constructors

Stack
Stacked

Group
Grouped

Overlay
Overlayed

Relative
Relative

data PlotConfig

Trace Independent Plot Config

Constructors

PlotConfig

Fields

title' :: !String

Plot Title (default "")

xlabel :: !String

X Axis Label (default "")

ylabel :: !String

Y Axis Label (default "")

xmode :: !AxisType

X Axis Type (default Linear)

ymode :: !AxisType

Y Axis Type (default Linear)

barMode :: !BarMode

Bar Mode (default Stack)

lineMode :: !Mode

Line Mode (default Linear)

colorScale :: !ColorScale

Color Scale for 3D (default Viridis)

legend :: !Bool

Show Legend (default True)

height :: !Int

Canvas Height in Pixels (default 800)

width :: !Int

Canvas Width in Pixels (default 800)

margin :: !Margin

Layout margin (default 0 0 0 0)

marker :: !Marker

Marker (default 0.5 Circle 1.0)

defaultConfig :: PlotConfig

Default PlotConfig

traceData :: Int -> Script

Generates line for data variable for N traces

traceLine :: ToJSON a => a -> Int -> ByteString

Generates line for ith trace variable

toScript :: ToJSON a => Maybe Layout -> [a] -> ByteString

Converts layout and list of traces to Script for insertion into HTML

mkTrace :: Maybe String -> Maybe Mode -> Maybe Marker -> Type -> [Double] -> [Double] -> [Double] -> Trace

Trace contructor, different order of arguments

mkTraceH :: ColorScale -> Bool -> Bool -> [String] -> [String] -> [[Double]] -> TraceH

Heatmap Trace constructor