Graphics.Plotly.Internal

Description

Internal Functions and Types for Plotly Plots in Haskell

Documentation

omitNulls :: [Pair] -> [Pair]

Thanks Alan Zimmerman https://gist.github.com/alanz/2465584

type Script = ByteString

Internal Type Alias

toHtml :: Script -> ByteString

Takes a Script and inserts it into the default HTML template

data Margin

Layout Margins

Constructors

Margin

Fields

l :: !Int

Left Margin

r :: !Int

Right Margin

b :: !Int

Bottom Margin

t :: !Int

Top Margin

data Layout

Layout for Plots

Constructors

Layout

Fields

title :: !String

Plot Title

xaxis :: !(Maybe Axis)

X Axis Layout

yaxis :: !(Maybe Axis)

Y Axis Layout

barmode :: !(Maybe BarMode)

Bar Chart Layout

height :: !Int

Plot Height in Pixels

width :: !Int

Plot Width in Pixels

showlegend :: !(Maybe Bool)

Show a Legend

margin :: !(Maybe Margin)

Margins

data AxisType

Axis Type

Constructors

Linear
Linear Axis (default)

Log
Logarithmic Axis

Date
Date Axis

Category
Categorical Axis

MultiCategory
Multi Category Axis (?)

data Axis

Axis Layout

Constructors

Axis

Fields

title :: !String

Axis Label

showGrid :: !Bool

Enable Grid

zeroLine :: !Bool

Enable Zero Line

type' :: !AxisType

Axis Type

data Line

Parallel Coordinate Lines

Constructors

Line

Fields

showscale :: !(Maybe Bool)

Show Color Bar

reversescale :: !(Maybe Bool)

Reverse the Color Scale

colorscale :: !(Maybe ColorScale)

Color Scale

color :: ![Double]

Values represnting the color

data Marker

Marker Config

Constructors

Marker

Fields

size :: !Double

Marker Size

symbol :: !Symbol

Marker Symbol

opacity :: !Double

Marker Opacity

data Trace

Data Trace

Constructors

Trace

Fields

name :: !(Maybe String)

Trace Name

x :: !(Maybe [Double])

X axis values

y :: !(Maybe [Double])

Y axis values

z :: !(Maybe [Double])

Z axis values

mode :: !(Maybe Mode)

Scatter Mode

type' :: !Type

Plot Type

barmode :: !(Maybe BarMode)

Histogram Bar Mode

xbins :: !(Maybe XBins)

Histogram Bins

marker :: !(Maybe Marker)

Marker Configuration

data TraceH

Heatmap Trace

Constructors

TraceH

Fields

z :: ![[Double]]

Data Matrix

x :: !(Maybe [String])

Y labels

y :: !(Maybe [String])

Y Labels

type' :: !Type

Heatmap

hoverOnGaps :: !(Maybe Bool)

Hover on Gaps

showScale :: !(Maybe Bool)

Show scale Bar

colorScale :: !(Maybe ColorScale)

Color Scale

data TraceS

Surface Trace

Constructors

TraceS

Fields

name :: !(Maybe String)

Trace Name

z :: ![[Double]]

Data Matrix

x :: !(Maybe [Double])

Y labels

y :: !(Maybe [Double])

Y Labels

type' :: !Type

Surface

colorScale :: !(Maybe ColorScale)

Color Scale

data TraceP

Parallel Coordinate Trace

Constructors

TraceP

Fields

type' :: !Type

ParCoords

line :: !(Maybe Line)

Line Configuration

dimensions :: ![Dimension]

Dimensions

data Dimension

Axis Dimension for Parallel Coordinate Plot

Constructors

Dimension

Fields

range :: !(Maybe [Int])

label :: !String

values :: ![Double]

data Type

Plot Type

Constructors

Scatter
Scatter Plot

Histogram
Histogram

Heatmap
Heatmap

Scatter3D
3D Scatter

Surface
3D Surface

ParCoords
Parallel Coordinate Plot

data Mode

Scatter Mode

Constructors

Lines
Draw Lines

Markers
Draw Markers only

LinesMarkers
Draw Both

data Symbol

Marker Symbols

Constructors

Circle

CircleOpen

Cross

Diamond

DiamondOpen

Square

SquareOpen

X

data ColorScale

Color scales for Heamap / 3D Plots

Constructors

Blackbody

Bluered

Blues

Cividis

Earth

Electric

Greens

Greys

Hot

Jet

Picnic

Portland

Rainbow

RdBu

Reds

Viridis

YlGnBu

YlOrRd

data BarMode

Bar Layout

Constructors

Stack
Stacked

Group
Grouped

Overlay
Overlayed

Relative
Relative

data XBins

Bins in Histogram

Constructors

XBins

Fields

size :: !Double

Size of Bins

start :: !(Maybe Double)

Start of Bins

end :: !(Maybe Double)

End of Bins

traceData :: Int -> Script

Generates line for data variable for N traces

traceLine :: ToJSON a => a -> Int -> ByteString

Generates line for ith trace variable

toScript :: ToJSON a => Maybe Layout -> [a] -> ByteString

Converts layout and list of traces to Script for insertion into HTML

mkTrace :: Maybe String -> Maybe Mode -> Maybe Marker -> Type -> BarMode -> XBins -> [Double] -> [Double] -> [Double] -> Trace

Trace contructor, different order of arguments

mkTraceH :: ColorScale -> Bool -> Bool -> [String] -> [String] -> [[Double]] -> TraceH

Heatmap Trace constructor

mkTraceS :: Maybe String -> ColorScale -> [Double] -> [Double] -> [[Double]] -> TraceS

Heatmap Trace constructor

mkTraceP :: ColorScale -> Bool -> Bool -> [Double] -> [String] -> [[Double]] -> TraceP

Parallel Coordinate Trace Constructor

mkTraceP' :: [String] -> [[Double]] -> TraceP

Parallel Coordinate Trace Constructor without Line configuration